Don't delay restart of sysstat "service", can break convergence
Currently, sysstat can break convergence when it attempts to restart, and fails
due to an exclusive lock on the active logfile held by sadc.

I think this is due to the more aggressive 10sec logging we're configuring sadc
to do. If the cron job runs for the first time before sysstat restarts, sadc
becomes active and holds the lock. This explains why we didn't see it in testing
-- there are more steps in other environments which opens up this race.

The sysstat service is a bit of a lie anyways - despite being a SysV init job,
it acts more like an upstart task than an actual service which daemonizes
itself.